+++ packages/apptive_grid_core/lib/apptive_grid_network.dart
@@ -6,20 +6,23 @@ import 'dart:convert';
 import 'package:apptive_grid_core/apptive_grid_model.dart';
 import 'package:apptive_grid_core/apptive_grid_options.dart';
 import 'package:apptive_grid_core/network/authentication/authentication_storage.dart';
+import 'package:apptive_grid_core/network/authentication/io_authenticator.dart'
+    if (dart.library.html) 'package:apptive_grid_core/network/authentication/web_authenticator.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:http/http.dart' as http;
 import 'package:openid_client/openid_client.dart';
 import 'package:uni_links/uni_links.dart' as uni_links;
-import 'package:apptive_grid_core/network/authentication/io_authenticator.dart'
-    if (dart.library.html) 'package:apptive_grid_core/network/authentication/web_authenticator.dart';
-
 import 'package:url_launcher/url_launcher.dart';
 
 export 'package:apptive_grid_core/network/authentication/web_auth_enabler/web_auth_enabler.dart';
 
 part 'network/apptive_grid_client.dart';
-part 'network/environment.dart';
+
 part 'network/authentication/apptive_grid_authentication_options.dart';
+
 part 'network/authentication/apptive_grid_authenticator.dart';
+
 part 'network/constants.dart';
+
+part 'network/environment.dart';