+++ packages/apptive_grid_core/lib/model/grid/grid_entry.dart
@@ -45,6 +45,9 @@ class GridEntry {
       case DataType.decimal:
         dataEntity = DecimalDataEntity(jsonData);
         break;
+      case DataType.attachment:
+        dataEntity = AttachmentDataEntity.fromJson(jsonData);
+        break;
     }
     return GridEntry(field, dataEntity);
   }