+++ packages/apptive_grid_form/lib/widgets/apptive_grid_form_widgets.dart
@@ -5,13 +5,18 @@ import 'dart:math';
 import 'package:apptive_grid_core/apptive_grid_core.dart';
 import 'package:apptive_grid_core/apptive_grid_model.dart';
 import 'package:apptive_grid_form/apptive_grid_form.dart';
+import 'package:apptive_grid_form/widgets/form_widget/attachment_manager.dart';
 import 'package:apptive_grid_form/widgets/form_widget/grid_row_dropdown/grid_row_dropdown_library.dart';
+import 'package:file_picker/file_picker.dart';
+import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:intl/intl.dart';
 import 'package:linked_scroll_controller/linked_scroll_controller.dart';
+import 'package:mime/mime.dart';
 
 part 'form_widget/form_widget.dart';
+part 'form_widget/attachment_form_widget.dart';
 part 'form_widget/check_box_form_widget.dart';
 part 'form_widget/enum_form_widget.dart';
 part 'form_widget/date_time_form_widget.dart';